Tidied up view imports and usings

+++ ProjectStorage.Web/Areas/Project/Controllers/FilesController.cs
@@ -1,19 +1,24 @@
 ﻿namespace ProjectStorage.Web.Areas.Project.Controllers
 {
+    using Constants;
+    using Data.Models;
     using Extensions;
+    using Microsoft.AspNetCore.Identity;
     using Microsoft.AspNetCore.Mvc;
     using Services;
     using System.Collections.Generic;
-    using System.Linq;  
+    using System.Linq;
 
     public class FilesController : ProjectBaseController
     {
         private readonly IList<string> textFileFormats;
         private readonly IFileService fileService;
+        private readonly UserManager<User> userManager;
 
-        public FilesController(IFileService fileService)
+        public FilesController(IFileService fileService, UserManager<User> userManager)
         {
             this.fileService = fileService;
+            this.userManager = userManager;
             this.textFileFormats = new List<string>
             {
                 "txt",
@@ -22,7 +27,11 @@
                 "cpp",
                 "rtf",
                 "csproj",
-                "sln"
+                "sln",
+                "html",
+                "css",
+                "cshtml",
+                "php"
             };
         }
 
@@ -36,7 +45,6 @@
             var extension = file.Path.Split('.').LastOrDefault();
             if (this.textFileFormats.Contains(extension))
             {
-         
                 return this.View("DetailsText", file);
             }
 
@@ -56,10 +64,14 @@
             {
                 return this.NotFound();
             }
-
+            if (!this.User.IsInRole(GlobalConstants.ProjectTesterRole) &&
+                !this.fileService.IsOwner(this.userManager.GetUserId(User), id))
+            {
+                return this.Redirect("/Account/Login");
+            }
             this.fileService.Delete(id);
 
-            return this.RedirectToAction("Index", "Manage", new { area = "Project" });
+            return this.RedirectToAction("Browse", "Manage", new { area = "Project" });
         }
 
         public IActionResult Download(string id)
